requirements-dev.txt
-r requirements.txt
django-debug-toolbar==4.0.0
django-extensions==3.2.0
ipython==8.10.0